Flytt vis observasjoner-knapp
Flyttes til home.page, siden den bare er nyttig der.
